JK: Added Content Negotiation: Accept: headers for GET Content-Type: headers for PUT Content-Type guessing either from the picture driver or from the file extension. The default value is text/html

+++ amaya/css.c
@@ -2526,7 +2526,7 @@ CSSInfoPtr          css;
 	sprintf (filename, "%s/%d/%s", TempFileDirectory, doc, file);
 	if (DumpCSSToFile (currentDocument, css, filename))
 	   return (-1);
-	res = PutObjectWWW (doc, filename, css->url, AMAYA_SYNC, (int) unknown_type,
+	res = PutObjectWWW (doc, filename, css->url, AMAYA_SYNC, unknown_type,
 			    (TTcbf *) NULL, (void *) NULL);
 	if (res != HT_OK)
 	  {
